Fix session history loading and the new-session workspace picker
The Android client opened conversations without tailTurns, so long
transcripts failed to decode or rendered empty. Match the web client:
load the last 120 turns, page older history, and decode u64 usage as Long.

New sessions hid the working directory behind a cramped top-level
dropdown. Show a searchable picker on the empty new-task screen, include
worktrees, and allow browsing the server filesystem.

Keep nested session history visible when a workspace is collapsed, pin
the transcript to its real bottom anchor, include child sessions in
search and activity, and persist a newly created conversation id.

testDebugUnitTest and assembleDebug are green. Version 1.2.1.

+++ app/src/main/java/app/codeg/android/feature/sessiondetail/timeline/TimelineNodeBody.kt
@@ -370,7 +370,7 @@ fun InlineTurnError(message: String, modifier: Modifier = Modifier) {
     }
 }
 
-private fun compactTokens(n: Int): String = when {
+private fun compactTokens(n: Long): String = when {
     n < 1_000 -> n.toString()
     n < 1_000_000 -> trimNum(n / 1_000.0) + "k"
     else -> trimNum(n / 1_000_000.0) + "M"
@@ -378,7 +378,7 @@ private fun compactTokens(n: Int): String = when {
 
 private fun trimNum(v: Double): String = if (v >= 100) v.roundToInt().toString() else String.format("%.1f", v)
 
-private fun compactDuration(ms: Int): String {
+private fun compactDuration(ms: Long): String {
     val seconds = ms / 1000.0
     if (seconds < 60) return String.format("%.1fs", seconds)
     val whole = seconds.roundToInt()
